+++ components/edit-collective/tiers/EditTierModal.js
@@ -7,31 +7,30 @@ import { omit } from 'lodash';
 import { FormattedMessage, useIntl } from 'react-intl';
 import styled from 'styled-components';
 
-import { CollectiveType } from '../../lib/constants/collectives';
-import { getGQLV2FrequencyFromInterval } from '../../lib/constants/intervals';
-import { AmountTypes, TierTypes } from '../../lib/constants/tiers-types';
-import { i18nGraphqlException } from '../../lib/errors';
-import { requireFields } from '../../lib/form-utils';
-import { API_V2_CONTEXT } from '../../lib/graphql/helpers';
-import { getCollectivePageRoute } from '../../lib/url-helpers';
+import { CollectiveType } from '../../../lib/constants/collectives';
+import { getGQLV2FrequencyFromInterval } from '../../../lib/constants/intervals';
+import { AmountTypes, TierTypes } from '../../../lib/constants/tiers-types';
+import { i18nGraphqlException } from '../../../lib/errors';
+import { requireFields } from '../../../lib/form-utils';
+import { API_V2_CONTEXT } from '../../../lib/graphql/helpers';
+import { getCollectivePageRoute } from '../../../lib/url-helpers';
 
-import { Flex } from '../Grid';
-import InputFieldPresets from '../InputFieldPresets';
-import InputSwitch from '../InputSwitch';
-import Link from '../Link';
-import StyledButton from '../StyledButton';
-import StyledHr from '../StyledHr';
-import StyledInput from '../StyledInput';
-import StyledInputAmount from '../StyledInputAmount';
-import StyledInputFormikField from '../StyledInputFormikField';
-import StyledLink from '../StyledLink';
-import StyledModal, { ModalBody, ModalFooter, ModalHeader } from '../StyledModal';
-import StyledSelect from '../StyledSelect';
-import StyledTextarea from '../StyledTextarea';
-import { P } from '../Text';
-import { TOAST_TYPE, useToasts } from '../ToastProvider';
-
-import ContributeTier from './ContributeTier';
+import ContributeTier from '../../contribute-cards/ContributeTier';
+import { Flex } from '../../Grid';
+import InputFieldPresets from '../../InputFieldPresets';
+import InputSwitch from '../../InputSwitch';
+import Link from '../../Link';
+import StyledButton from '../../StyledButton';
+import StyledHr from '../../StyledHr';
+import StyledInput from '../../StyledInput';
+import StyledInputAmount from '../../StyledInputAmount';
+import StyledInputFormikField from '../../StyledInputFormikField';
+import StyledLink from '../../StyledLink';
+import StyledModal, { ModalBody, ModalFooter, ModalHeader } from '../../StyledModal';
+import StyledSelect from '../../StyledSelect';
+import StyledTextarea from '../../StyledTextarea';
+import { P } from '../../Text';
+import { TOAST_TYPE, useToasts } from '../../ToastProvider';
 
 const { FUND, PROJECT } = CollectiveType;
 const { TIER, TICKET, MEMBERSHIP, SERVICE, PRODUCT, DONATION } = TierTypes;
