+++ stories/collective-page/EditTierModal.stories.mdx
@@ -1,5 +1,5 @@
 import { ArgsTable, Meta, Story } from '@storybook/addon-docs/blocks';
-import EditTierModal from '../../components/contribute-cards/EditTierModal';
+import EditTierModal from '../../components/edit-collective/tiers/EditTierModal';
 
 <Meta title="Collective Page/EditTierModal" component={EditTierModal} argTypes={{}} />
